itmo_tg_bot_final/requirements.txt
beautifulsoup4>=4.10.0
requests>=2.25.1
python-dotenv>=0.19.0
python-telegram-bot>=20.0
openai>=0.27.0
faiss-cpu>=1.7.3
sentence-transformers>=2.2.2
PyPDF2>=3.0.0
python-certifi-win32>=0.0.0
selenium>=4.8.0